consolidate state for dashboard
