+++ front-end/src/dashboard/MyPie.js
@@ -1,13 +1,10 @@
-// const { useEffect, useState } = require("react");
 import React, { useEffect, useState } from "react";
-import { payLoad } from "../constants/index";
 import { Pie } from "react-chartjs-2";
 
-export const MyPie = () => {
+export const MyPie = ({ stats }) => {
   const [average, setAverage] = useState(0);
   const [state, setState] = useState({
     labels: [
-      // "Promedio de Pruebas",
       "# Preguntas Correctas",
       "# Preguntas Incorrectas",
     ],
@@ -25,48 +22,39 @@ export const MyPie = () => {
   });
 
   useEffect(() => {
-    const payLoad = {
-      method: "GET",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${localStorage.getItem('token')}`,
-      },
-    };
-    fetch("http://localhost:3000/total", payLoad)
-      .then((r) => r.json())
-      .then((statsObj) => {
-        setState((prevState) => {
-          return {
-            ...prevState,
-            datasets: [
-              {
-                label: "Resultados de Pruebas",
-                backgroundColor: [
-                  // "#C9DE00",
-                  "#fce38a",
-                  "#f38181",
-                  "#00A6B4",
-                  "#6800B4",
-                ],
-                hoverBackgroundColor: [
-                  "#501800",
-                  "#4B5000",
-                  "#175000",
-                  "#003350",
-                  "#35014F",
-                ],
-                data: [
-                  // statsObj.test_average,
-                  statsObj.number_correct,
-                  statsObj.number_incorrect,
-                ],
-              },
+    setStats(stats)
+  }, [stats]);
+
+  const setStats = (stats) => {
+    setState((prevState) => {
+      return {
+        ...prevState,
+        datasets: [
+          {
+            label: "Resultados de Pruebas",
+            backgroundColor: [
+              "#fce38a",
+              "#f38181",
+              "#00A6B4",
+              "#6800B4",
             ],
-          };
-        });
-        setAverage(statsObj.test_average);
-      });
-  }, []);
+            hoverBackgroundColor: [
+              "#501800",
+              "#4B5000",
+              "#175000",
+              "#003350",
+              "#35014F",
+            ],
+            data: [
+              stats.number_correct,
+              stats.number_incorrect,
+            ],
+          },
+        ],
+      };
+    });
+    setAverage(stats.test_average);
+  }
   return (
     <div>
       <Pie
